Split Player and Item's networking and client things
This is so devs can access client sided things without needing the server API.
