+++ SigurdLib.Common/Features/Patches/LocalLeftPatch.cs
@@ -1,0 +1,17 @@
+using GameNetcodeStuff;
+using HarmonyLib;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Sigurd.Common.Features.Patches
+{
+    [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.OnLocalDisconnect))]
+    internal static class LocalLeft
+    {
+        private static void Prefix()
+        {
+            Player.Dictionary.Clear();
+        }
+    }
+}
