+++ SigurdLib.ServerAPI/Events/EventArgs/Player/JoinedEventArgs.cs
@@ -1,20 +1,20 @@
 ﻿namespace Sigurd.ServerAPI.Events.EventArgs.Player
 {
     /// <summary>
-    /// Contains all the information after a <see cref="Features.Player"/> joined the server. Including the host.
+    /// Contains all the information after a <see cref="Common.Features.Player"/> joined the server. Including the host.
     /// </summary>
     public class JoinedEventArgs : System.EventArgs
     {
         /// <summary>
         /// Gets the joined player.
         /// </summary>
-        public Features.Player Player { get; }
+        public Common.Features.Player Player { get; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="JoinedEventArgs"/> class.
         /// </summary>
         /// <param name="player"><inheritdoc cref="Player" /></param>
-        public JoinedEventArgs(Features.Player player)
+        public JoinedEventArgs(Common.Features.Player player)
         {
             Player = player;
         }
